Chapter 6.2 - Generate and verify signed JWT tokens.
Introduces new SignedJwtTokenStore class which is an implementation of TokenStore.
Unlike DatabaseTokenStore, it doesn't store anything in a persistent storage.
Instead, it only creates a signed version of the Token
and then it verifies it.

It uses the same macKey as HmacTokenStore - notice the usage of MACSigner and MACVerifier classes.

Also notice how the read method checks the audience.