Chapter 7.4.1 Add OAuth2TokenStore.
This implements a new token store using OAuth2 tokens.

It uses "introspection endpoint" which is in our case ForgeRock Access Management product
running in a docker container on the same machine (see Appendinx A for instructions).
We use the introspection endpoint to validate tokens supplied by the client.
The endpoint should respond with 200 and "active" field _true_ if the token is valid.

If that holds, we extract some more information from the token such as "exp", "sub", "scope", and "client_id".
Note that all these fields are optional (see Table 7.1 on p. 242) and they might be missing in some OAuth2 AS implementations.