Chapter 7.4.4 SignedJwtAccessTokenStore.
The public key for verification is retrieved from `jwkSetUri`.
The key is then used to validate the signature of the token (see the `read` method).
In that process, `signatureAlgorithm` is used.

Separately, we verify the issuer and the audience - this is important to prevent certain types of attacks (see Ch6).

Finally, we parse token scopes - depending on the AS implementation, we support either single string in the "scope" field,
or a list of strings.