Chapter 8.1 Adding groups
UserController is modified to set and use groups when checking permissions.
This is doen by using the previously defined `permissions` view.

WARNING: here we only set request attribute "groups" for basic authentication.
TokenController, at the moment, only sets "subject" so in UserController#requirePermissions
we need to check if "groups" is actually different from null,
otherwise the app fails with NPE.