Ch 9.2.3 Capability URIs for messages inside a space.
With only a space URI, the client cannot do much - we don't even have a resource
for showing space details.
There are messages, however, the client cannot access the
since it cannot easily construct the uri by appending '/messages'.

In the spirit of HATEOS, we need to include all the relevant links in the response!

We actually return 3 different messages links:
- "messages" for full access
- "messages-rw" for read-write (no delete) access
- "messages-r" for read-only access